% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanciones.R
\name{beneficiarios}
\alias{beneficiarios}
\title{Beneficiaries}
\usage{
beneficiarios(busqueda, ambitos = "G", ...)
}
\arguments{
\item{busqueda}{At least three characters of the entity you are searching.}

\item{ambitos}{Organization level to search for it.}

\item{...}{Other queries.}
}
\value{
A data.frame with the id, description, NIF and name of the entity.
}
\description{
Look for entities that received a grant.
}
\examples{
out <- beneficiarios("123")
}
