% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organos.R
\name{concesiones}
\alias{concesiones}
\title{Title}
\usage{
concesiones()
}
\description{
Title
}
\references{
\url{https://www.pap.hacienda.gob.es/bdnstrans/GE/es/concesiones/consulta}
}
\seealso{
\code{\link[=de_minimis]{de_minimis()}}, \code{\link[=grandes_beneficiados]{grandes_beneficiados()}}, \code{\link[=political_parties]{political_parties()}}, \code{\link[=ayudas_estado]{ayudas_estado()}}
}
