% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{configuration}
\alias{configuration}
\title{Site configuration}
\usage{
configuration()
}
\value{
A list with the date of the update, legal warning and allowed routes.
}
\description{
Check the website configuration.
}
\examples{
out <- configuration()
out$ultimaActualizacion
}
