% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentos.R
\name{documento}
\alias{documento}
\title{Download document}
\usage{
documento(document, name)
}
\arguments{
\item{document}{Id of the document to download.}

\item{name}{Path and name of the document on your system.}
}
\value{
The path of the file if it has been successfully saved.
}
\description{
Download document
}
\examples{
documento(963765, "~/Downloads/document_963765.pdf")
}
