% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organos.R
\name{organism}
\alias{organism}
\title{Organism}
\usage{
organism(admin = "C", ...)
}
\arguments{
\item{admin}{Level of the organization: one of \code{c(Estatal = "C", Local = "L", Autonomica = "A", Otros = "O")}.}

\item{...}{Other parameters to search them.}
}
\value{
A data.frame with the description, id, parent_id and admin level.
}
\description{
Look up for public organism.
}
\examples{
oc <- organism("C")
ol <- organism("L")
oa <- organism("A")
oo <- organism("O")
}
