% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convocatorias.R
\name{search_convocatorias}
\alias{search_convocatorias}
\title{Search calls}
\usage{
search_convocatorias(n = 100, fecha_inicio, fecha_fin, ...)
}
\arguments{
\item{...}{}
}
\description{
Looks up which calls are there.
}
\references{
\url{https://www.pap.hacienda.gob.es/bdnstrans/GE/es/convocatorias}
}
