% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planes_estrategicos.R
\name{strategic_plans}
\alias{strategic_plans}
\title{Search strategic plans}
\usage{
strategic_plans(...)
}
\arguments{
\item{...}{}
}
\description{
Search strategic plans
}
\references{
\url{https://www.pap.hacienda.gob.es/bdnstrans/GE/es/planesEstrategicos}
}
